requirements.txt
Flask
PyMuPDF
python-docx
openai
sentence-transformers
scikit-learn
python-dotenv
spacy
flask-cors
